requirements.txt
django>=1.11.6
sqlalchemy>=1.1.14
celery>=4.1
django-celery-beat>=1.0.0
djangorestframework>=3.7.0
markdown>=2.6.9
django-filter>=1.0.4
psycopg2>=2.7.3